+++ .size-limit.js
@@ -4,4 +4,19 @@ module.exports = [
     // ugly
     limit: '15000ms',
   },
+  {
+    path: ['webapp/public/assets/app.css'],
+  },
+  {
+    path: ['webapp/public/assets/styles.css'],
+  },
+  {
+    path: ['packages/pyroscope-flamegraph/dist/index.js'],
+  },
+  {
+    path: ['packages/pyroscope-flamegraph/dist/index.node.js'],
+  },
+  {
+    path: ['packages/pyroscope-flamegraph/dist/index.css'],
+  },
 ];