+++ packages/pyroscope-flamegraph/src/FlameGraph/FlameGraphComponent/index.tsx
@@ -1,11 +1,16 @@
 /* eslint-disable no-unused-expressions */
 import React, { useCallback, useRef } from 'react';
 import clsx from 'clsx';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faRedo } from '@fortawesome/free-solid-svg-icons/faRedo';
+import { faCopy } from '@fortawesome/free-solid-svg-icons/faCopy';
+import { faHighlighter } from '@fortawesome/free-solid-svg-icons/faHighlighter';
+import { faCompressAlt } from '@fortawesome/free-solid-svg-icons/faCompressAlt';
 import { MenuItem } from '@szhsin/react-menu';
 import useResizeObserver from '@react-hook/resize-observer';
 import { Maybe } from 'true-myth';
 import debounce from 'lodash.debounce';
-import { Flamebearer } from '@pyroscope/models';
+import { Flamebearer } from '@pyroscope/models/src';
 import styles from './canvas.module.css';
 import Flamegraph from './Flamegraph';
 import Highlight from './Highlight';
@@ -25,9 +30,11 @@ interface FlamegraphProps {
   highlightQuery: ConstructorParameters<typeof Flamegraph>[4];
   zoom: ConstructorParameters<typeof Flamegraph>[5];
   showCredit: boolean;
+  selectedItem: Maybe<string>;
 
   onZoom: (bar: Maybe<{ i: number; j: number }>) => void;
   onFocusOnNode: (i: number, j: number) => void;
+  setActiveItem: (item: { name: string }) => void;
 
   onReset: () => void;
   isDirty: () => boolean;
@@ -56,6 +63,8 @@ export default function FlameGraphComponent(props: FlamegraphProps) {
     zoom,
     toolbarVisible,
     showCredit,
+    setActiveItem,
+    selectedItem,
   } = props;
 
   const { onZoom, onReset, isDirty, onFocusOnNode } = props;
@@ -144,8 +153,9 @@ export default function FlameGraphComponent(props: FlamegraphProps) {
     (x: number, y: number) => {
       const dirty = isDirty();
       const bar = getFlamegraph().xyToBar(x, y);
+      const barName = bar.isJust ? bar.value.name : '';
 
-      const FocusItem = () => {
+      const CollapseItem = () => {
         const hoveredOnValidNode = bar.mapOrElse(
           () => false,
           () => true
@@ -162,19 +172,56 @@ export default function FlameGraphComponent(props: FlamegraphProps) {
             disabled={!hoveredOnValidNode}
             onClick={onClick}
           >
-            Focus on this subtree
+            <FontAwesomeIcon icon={faCompressAlt} />
+            Collapse nodes above
+          </MenuItem>
+        );
+      };
+
+      const CopyItem = () => {
+        const onClick = () => {
+          if (!navigator.clipboard) return;
+
+          navigator.clipboard.writeText(barName);
+        };
+
+        return (
+          <MenuItem key="copy" onClick={onClick}>
+            <FontAwesomeIcon icon={faCopy} />
+            Copy function name
+          </MenuItem>
+        );
+      };
+
+      const HighlightSimilarNodesItem = () => {
+        const onClick = () => {
+          setActiveItem({ name: barName });
+        };
+
+        const actionName =
+          selectedItem.isJust && selectedItem.value === barName
+            ? 'Clear highlight'
+            : 'Highlight similar nodes';
+
+        return (
+          <MenuItem key="highlight-similar-nodes" onClick={onClick}>
+            <FontAwesomeIcon icon={faHighlighter} />
+            {actionName}
           </MenuItem>
         );
       };
 
       return [
         <MenuItem key="reset" disabled={!dirty} onClick={onReset}>
+          <FontAwesomeIcon icon={faRedo} />
           Reset View
         </MenuItem>,
-        FocusItem(),
+        CollapseItem(),
+        CopyItem(),
+        HighlightSimilarNodesItem(),
       ];
     },
-    [flamegraph]
+    [flamegraph, selectedItem]
   );
 
   const constructCanvas = () => {
@@ -248,7 +295,6 @@ export default function FlameGraphComponent(props: FlamegraphProps) {
         setPalette={setPalette}
         toolbarVisible={toolbarVisible}
       />
-
       <div
         data-testid={dataTestId}
         style={{