+++ pkg/server/controller.go
@@ -9,6 +9,7 @@ import (
 	"net/http"
 	"net/http/pprof"
 	"net/url"
+	"os"
 	"path/filepath"
 	"sync"
 	"sync/atomic"
@@ -415,7 +416,10 @@ func (ctrl *Controller) getHandler() (http.Handler, error) {
 		return nil, err
 	}
 
-	return ctrl.corsMiddleware()(gzhttpMiddleware(handler)), nil
+	h := ctrl.corsMiddleware()(gzhttpMiddleware(handler))
+	h = ctrl.logginMiddleware(h)
+
+	return h, nil
 }
 
 func (ctrl *Controller) Start() error {
@@ -546,4 +550,14 @@ func expectFormats(format string) error {
 	default:
 		return errUnknownFormat
 	}
+}
+
+func (ctrl *Controller) logginMiddleware(next http.Handler) http.Handler {
+	if ctrl.config.LogLevel == "debug" {
+		// log to Stdout using Apache Common Log Format
+		// TODO maybe use JSON?
+		return handlers.LoggingHandler(os.Stdout, next)
+	}
+
+	return next
 }