+++ webapp/CHANGELOG.md
@@ -3,6 +3,132 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.21.11](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.21.10...@pyroscope/webapp@1.21.11) (2022-07-20)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
+## [1.21.10](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.21.9...@pyroscope/webapp@1.21.10) (2022-07-20)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
+## [1.21.9](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.21.8...@pyroscope/webapp@1.21.9) (2022-07-20)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
+## [1.21.8](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.21.7...@pyroscope/webapp@1.21.8) (2022-07-19)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
+## [1.21.7](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.21.6...@pyroscope/webapp@1.21.7) (2022-07-19)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
+## [1.21.6](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.21.5...@pyroscope/webapp@1.21.6) (2022-07-18)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
+## [1.21.5](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.21.4...@pyroscope/webapp@1.21.5) (2022-07-18)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
+## [1.21.4](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.21.3...@pyroscope/webapp@1.21.4) (2022-07-18)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
+## [1.21.3](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.21.2...@pyroscope/webapp@1.21.3) (2022-07-15)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
+## [1.21.2](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.21.1...@pyroscope/webapp@1.21.2) (2022-07-15)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
+## [1.21.1](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.21.0...@pyroscope/webapp@1.21.1) (2022-07-15)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
+# [1.21.0](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.20.0...@pyroscope/webapp@1.21.0) (2022-07-15)
+
+
+### Features
+
+* update right-click context menu ([#1259](https://github.com/pyroscope-io/pyroscope/issues/1259)) ([8aea02f](https://github.com/pyroscope-io/pyroscope/commit/8aea02f56320daacfd753d73db6936dcc7cdaef8))
+
+
+
+
+
+# [1.20.0](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.19.3...@pyroscope/webapp@1.20.0) (2022-07-13)
+
+
+### Features
+
+* add new tooltip design ([#1246](https://github.com/pyroscope-io/pyroscope/issues/1246)) ([8345168](https://github.com/pyroscope-io/pyroscope/commit/83451683d131671771b0e97e052068b08bfe35bd))
+
+
+
+
+
+## [1.19.3](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.19.2...@pyroscope/webapp@1.19.3) (2022-07-13)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
+## [1.19.2](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.19.1...@pyroscope/webapp@1.19.2) (2022-07-13)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
 ## [1.19.1](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.19.0...@pyroscope/webapp@1.19.1) (2022-07-13)
 
 **Note:** Version bump only for package @pyroscope/webapp